ppdiffusers/examples/class_conditional_image_generation/requirements.txt
paddlenlp>=2.7.2
ppdiffusers>=0.19.3
fastcore
visualdl
Pillow